+++ pypose/utils/__init__.py
@@ -1,1 +1,1 @@
-from .planner import ReduceToBason
+from .stepper import ReduceToBason
